chore(backend): run pagination tests via shared subresource tests
Create fewer instances in shared pagination tests.
